TVG-102 Blueprint for frontend routes
